fix(TCOMP-911): move all Validation classes to validation package
