+++ main/plugins/org.talend.sdk.component.studio-integration/src/main/java/org/talend/sdk/component/studio/model/parameter/listener/MinValidator.java
@@ -16,7 +16,7 @@
 package org.talend.sdk.component.studio.model.parameter.listener;
 
 import org.talend.sdk.component.studio.i18n.Messages;
-import org.talend.sdk.component.studio.model.parameter.ValidationLabel;
+import org.talend.sdk.component.studio.model.parameter.validation.ValidationLabel;
 
 /**
  * Checks whether new value is greater than specified {@code min} value
